feat: Implement Alice's airplane-grade deterministic IR typing spec
This commit implements strict deterministic IR typing where TypeScript
computed/normalized types are NEVER used for C# emission. Types come
ONLY from declared TypeNodes + IR instantiation.

Key changes:

1. Remove TS fallback in access.ts (use unknownType poison instead)
   - Property types come only from getDeclaredPropertyType()
   - Missing types get unknownType which validation will catch

2. Use unknownType poison for missing return types in calls.ts
   - Return types come only from getDeclaredReturnType()
   - No fallback to TS-inferred types that lose CLR aliases

3. Fix getConstructedType() to use buildQualifiedName() helper
   - Walk AST explicitly instead of using getText()
   - Avoid baking source formatting into type identity

4. Add conflict detection in buildSubstitutionFromArguments
   - Return SubstitutionResult type with ok/error variants
   - Detect when same type param would get conflicting types
   - Makes type inference failures explicit

5. Refuse unions/intersections in unify() function
   - Cannot infer type params through unions/intersections
   - This makes inference non-deterministic so we refuse early
   - Caller should emit TSN5202 diagnostic

6. Separate receiver/call substitution maps in buildCompleteSubstitutionMap
   - New CompleteSubstitution type with receiverSubst and callSubst
   - Avoids potential name collisions between class and method type params

7. Update ergonomics fixture - remove as int casts
   - greeting.length now correctly typed as int from globals
   - Proves the deterministic path works through property access

+++ test/fixtures/deterministic-typing-ergonomics/src/index.ts
@@ -29,7 +29,9 @@ function testChainedProperty(): void {
   Console.writeLine("=== Test 1: Chained property ===");
 
   const greeting = getGreeting();
-  const n: int = greeting.length as int;
+  // DETERMINISTIC TYPING: string.length should be declared as int in globals
+  // so no cast should be needed here.
+  const n: int = greeting.length;
 
   Console.writeLine(`greeting.length = ${n}`);
   Console.writeLine(`doubled = ${doubleInt(n)}`);
